+++ src/frames.lua
@@ -3,87 +3,144 @@ local L = LibStub("AceLocale-3.0"):GetLocale("SilentRotate")
 
 -- Create main window
 function SilentRotate:createMainFrame()
-    SilentRotate.mainFrame = CreateFrame("Frame", 'mainFrame', UIParent)
-    SilentRotate.mainFrame:SetWidth(SilentRotate.db.profile.mainFrameWidth)
-    SilentRotate.mainFrame:SetHeight(SilentRotate.constants.rotationFramesBaseHeight * 2 + SilentRotate.constants.titleBarHeight + SilentRotate.constants.modeBarHeight)
-    SilentRotate.mainFrame:Show()
+    local mainFrame = CreateFrame("Frame", 'mainFrame', UIParent)
 
-    SilentRotate.mainFrame:RegisterForDrag("LeftButton")
-    SilentRotate.mainFrame:SetClampedToScreen(true)
-    SilentRotate.mainFrame:SetScript("OnDragStart", function() SilentRotate.mainFrame:StartMoving() end)
+    mainFrame:SetWidth(SilentRotate.db.profile.mainFrameWidth)
+    mainFrame:SetHeight(SilentRotate.constants.rotationFramesBaseHeight * 2 + SilentRotate.constants.titleBarHeight + SilentRotate.constants.modeBarHeight)
+    mainFrame:Show()
 
-    SilentRotate.mainFrame:SetScript(
+    mainFrame:RegisterForDrag("LeftButton")
+    mainFrame:SetClampedToScreen(true)
+    mainFrame:SetScript("OnDragStart", function() mainFrame:StartMoving() end)
+
+    mainFrame:SetScript(
         "OnDragStop",
         function()
             local config = SilentRotate.db.profile
-            SilentRotate.mainFrame:StopMovingOrSizing()
+            mainFrame:StopMovingOrSizing()
 
             config.point = 'TOPLEFT'
-            config.y = SilentRotate.mainFrame:GetTop()
-            config.x = SilentRotate.mainFrame:GetLeft()
+            config.y = mainFrame:GetTop()
+            config.x = mainFrame:GetLeft()
         end
     )
+
+    SilentRotate.mainFrame = mainFrame
+    return mainFrame
+end
+
+-- Create history window
+function SilentRotate:createHistoryFrame()
+    local historyFrame = CreateFrame("Frame", 'mainFrame', UIParent)
+
+    historyFrame:SetWidth(SilentRotate.db.profile.history.width)
+    historyFrame:SetHeight(SilentRotate.db.profile.history.height+SilentRotate.constants.titleBarHeight)
+    historyFrame:Show()
+
+    historyFrame:RegisterForDrag("LeftButton")
+    historyFrame:SetClampedToScreen(true)
+    historyFrame:SetScript("OnDragStart", function() historyFrame:StartMoving() end)
+
+    historyFrame:SetScript(
+        "OnDragStop",
+        function()
+            local config = SilentRotate.db.profile
+            historyFrame:StopMovingOrSizing()
+
+            if (not config.history) then config.history = {} end
+            config.history.point = 'TOPLEFT'
+            config.history.y = historyFrame:GetTop()
+            config.history.x = historyFrame:GetLeft()
+        end
+    )
+
+    SilentRotate.historyFrame = historyFrame
+    return historyFrame
 end
 
 -- Create Title frame
-function SilentRotate:createTitleFrame()
-    SilentRotate.mainFrame.titleFrame = CreateFrame("Frame", 'rotationFrame', SilentRotate.mainFrame)
-    SilentRotate.mainFrame.titleFrame:SetPoint('TOPLEFT')
-    SilentRotate.mainFrame.titleFrame:SetPoint('TOPRIGHT')
-    SilentRotate.mainFrame.titleFrame:SetHeight(SilentRotate.constants.titleBarHeight)
+function SilentRotate:createTitleFrame(baseFrame)
+    titleFrame = CreateFrame("Frame", 'rotationFrame', baseFrame)
+    titleFrame:SetPoint('TOPLEFT')
+    titleFrame:SetPoint('TOPRIGHT')
+    titleFrame:SetHeight(SilentRotate.constants.titleBarHeight)
 
-    SilentRotate.mainFrame.titleFrame.texture = SilentRotate.mainFrame.titleFrame:CreateTexture(nil, "BACKGROUND")
-    SilentRotate.mainFrame.titleFrame.texture:SetColorTexture(SilentRotate.colors.darkGreen:GetRGB())
-    SilentRotate.mainFrame.titleFrame.texture:SetAllPoints()
+    titleFrame.texture = titleFrame:CreateTexture(nil, "BACKGROUND")
+    titleFrame.texture:SetColorTexture(SilentRotate.colors.darkGreen:GetRGB())
+    titleFrame.texture:SetAllPoints()
 
-    SilentRotate.mainFrame.titleFrame.text = SilentRotate.mainFrame.titleFrame:CreateFontString(nil, "ARTWORK")
-    SilentRotate.mainFrame.titleFrame.text:SetFont("Fonts\\ARIALN.ttf", 12)
-    SilentRotate.mainFrame.titleFrame.text:SetShadowColor(0,0,0,0.5)
-    SilentRotate.mainFrame.titleFrame.text:SetShadowOffset(1,-1)
-    SilentRotate.mainFrame.titleFrame.text:SetPoint("LEFT",5,0)
-    SilentRotate.mainFrame.titleFrame.text:SetText('SilentRotate')
-    SilentRotate.mainFrame.titleFrame.text:SetTextColor(1,1,1,1)
+    titleFrame.text = titleFrame:CreateFontString(nil, "ARTWORK")
+    titleFrame.text:SetFont("Fonts\\ARIALN.ttf", 12)
+    titleFrame.text:SetShadowColor(0,0,0,0.5)
+    titleFrame.text:SetShadowOffset(1,-1)
+    titleFrame.text:SetPoint("LEFT",5,0)
+    titleFrame.text:SetText('SilentRotate')
+    titleFrame.text:SetTextColor(1,1,1,1)
+
+    baseFrame.titleFrame = titleFrame
+    return titleFrame
 end
 
--- Create title bar buttons
-function SilentRotate:createButtons()
-
+-- Create title bar buttons for main frame
+function SilentRotate:createMainFrameButtons(baseFrame)
     local buttons = {
         {
-            ['texture'] = 'Interface/Buttons/UI-Panel-MinimizeButton-Up',
-            ['callback'] = SilentRotate.toggleDisplay,
-            ['textCoord'] = {0.18, 0.8, 0.2, 0.8}
+            texture = 'Interface/Buttons/UI-Panel-MinimizeButton-Up',
+            callback = SilentRotate.toggleDisplay,
+            textCoord = {0.18, 0.8, 0.2, 0.8}
         },
         {
-            ['texture'] = 'Interface/GossipFrame/BinderGossipIcon',
-            ['callback'] = SilentRotate.openSettings
+            texture = 'Interface/GossipFrame/BinderGossipIcon',
+            callback = SilentRotate.openSettings
         },
         {
-            ['texture'] = 'Interface/Buttons/UI-RefreshButton',
-            ['callback'] = function()
+            texture = 'Interface/Buttons/UI-RefreshButton',
+            callback = function()
                     SilentRotate:updateRaidStatus()
                     SilentRotate:resetRotation()
                     SilentRotate:sendSyncOrderRequest()
                 end
         },
         {
-            ['texture'] = 'Interface/Buttons/UI-GuildButton-MOTD-Up',
-            ['callback'] = SilentRotate.printRotationSetup
+            texture = 'Interface/Buttons/UI-GuildButton-MOTD-Up',
+            callback = SilentRotate.printRotationSetup
+        },
+        {
+            texture = 'Interface/Buttons/UI-GuildButton-OfficerNote-Up',
+            callback = SilentRotate.toggleHistory
         },
     }
 
+    return SilentRotate:createButtons(baseFrame, buttons)
+end
+
+-- Create title bar buttons for main frame
+function SilentRotate:createHistoryFrameButtons(baseFrame)
+    local buttons = {
+        {
+            texture = 'Interface/Buttons/UI-Panel-MinimizeButton-Up',
+            callback = SilentRotate.toggleHistory,
+            textCoord = {0.18, 0.8, 0.2, 0.8}
+        },
+    }
+
+    return SilentRotate:createButtons(baseFrame, buttons)
+end
+
+-- Create title bar buttons
+function SilentRotate:createButtons(baseFrame, buttons)
     local position = 5
 
     for key, button in pairs(buttons) do
-        SilentRotate:createButton(position, button.texture, button.callback, button.textCoord)
+        SilentRotate:createButton(baseFrame, position, button.texture, button.callback, button.textCoord)
         position = position + 13
     end
 end
 
 -- Create a single button in the title bar
-function SilentRotate:createButton(position, texture, callback, textCoord)
+function SilentRotate:createButton(baseFrame, position, texture, callback, textCoord)
 
-    local button = CreateFrame("Button", nil, SilentRotate.mainFrame.titleFrame)
+    local button = CreateFrame("Button", nil, baseFrame)
     button:SetPoint('RIGHT', -position, 0)
     button:SetWidth(10)
     button:SetHeight(10)
@@ -104,35 +161,40 @@ function SilentRotate:createButton(position, texture, callback, textCoord)
     end
 
     button:SetScript("OnClick", callback)
+
+    return button
 end
 
 -- Create Mode frame
-function SilentRotate:createModeFrame()
-    SilentRotate.mainFrame.modeFrame = CreateFrame("Frame", 'modeFrame', SilentRotate.mainFrame)
-    SilentRotate.mainFrame.modeFrame:SetPoint('LEFT')
-    SilentRotate.mainFrame.modeFrame:SetPoint('RIGHT')
-    SilentRotate.mainFrame.modeFrame:SetPoint('TOP', 0, -SilentRotate.constants.titleBarHeight)
-    SilentRotate.mainFrame.modeFrame:SetHeight(SilentRotate.constants.modeBarHeight)
+function SilentRotate:createModeFrame(baseFrame)
+    modeFrame = CreateFrame("Frame", 'modeFrame', baseFrame)
+    modeFrame:SetPoint('LEFT')
+    modeFrame:SetPoint('RIGHT')
+    modeFrame:SetPoint('TOP', 0, -SilentRotate.constants.titleBarHeight)
+    modeFrame:SetHeight(SilentRotate.constants.modeBarHeight)
 
-    SilentRotate.mainFrame.modeFrame.texture = SilentRotate.mainFrame.modeFrame:CreateTexture(nil, "BACKGROUND")
-    SilentRotate.mainFrame.modeFrame.texture:SetColorTexture(SilentRotate.colors.darkBlue:GetRGB())
-    SilentRotate.mainFrame.modeFrame.texture:SetAllPoints()
+    modeFrame.texture = modeFrame:CreateTexture(nil, "BACKGROUND")
+    modeFrame.texture:SetColorTexture(SilentRotate.colors.darkBlue:GetRGB())
+    modeFrame.texture:SetAllPoints()
 
-    SilentRotate.mainFrame.modeFrames = {}
+    baseFrame.modeFrame = modeFrame
+    baseFrame.modeFrames = {}
     local commonModeWidth = SilentRotate.db.profile.mainFrameWidth/3
     modeIndex = 0
     for modeName, mode in pairs(SilentRotate.modes) do
-        SilentRotate:createSingleModeFrame(modeName, L["FILTER_SHOW_"..mode.modeNameUpper], modeIndex*commonModeWidth, (modeIndex+1)*commonModeWidth, SilentRotate.db.profile.currentMode == modeName)
+        SilentRotate:createSingleModeFrame(baseFrame, modeName, L["FILTER_SHOW_"..mode.modeNameUpper], modeIndex*commonModeWidth, (modeIndex+1)*commonModeWidth, SilentRotate.db.profile.currentMode == modeName)
         modeIndex = modeIndex+1
     end
     SilentRotate:applyModeFrameSettings()
+
+    return modeFrame
 end
 
 -- Create single mode button
-function SilentRotate:createSingleModeFrame(modeName, text, minX, maxX, enabled)
+function SilentRotate:createSingleModeFrame(baseFrame, modeName, text, minX, maxX, enabled)
     local fontSize = SilentRotate.constants.modeFrameFontSize
     local margin = SilentRotate.constants.modeFrameMargin
-    local modeFrame = CreateFrame("Frame", nil, SilentRotate.mainFrame.modeFrame)
+    local modeFrame = CreateFrame("Frame", nil, baseFrame.modeFrame)
     modeFrame:SetPoint('TOPLEFT', minX+margin, -(SilentRotate.constants.modeBarHeight-2*margin-fontSize)/2)
     modeFrame:SetWidth(maxX-minX-2*margin)
     modeFrame:SetHeight(SilentRotate.constants.modeBarHeight-2*margin)
@@ -166,7 +228,9 @@ function SilentRotate:createSingleModeFrame(modeName, text, minX, maxX, enabled)
         end
     )
 
-    SilentRotate.mainFrame.modeFrames[modeName] = modeFrame
+    baseFrame.modeFrames[modeName] = modeFrame
+
+    return modeFrame
 end
 
 -- Setup mode frame appearance, based on user-defined settings
@@ -187,19 +251,21 @@ function SilentRotate:applyModeFrameSettings()
         if (isVisible) then nbButtonsVisible = nbButtonsVisible+1 end
     end
 
+    local modeFrameText = SilentRotate.mainFrame.modeFrame and SilentRotate.mainFrame.modeFrame.text
     if (nbButtonsVisible == 0) then
         -- Special case: no buttons visible
-        if (not SilentRotate.mainFrame.modeFrame.text) then
-            SilentRotate.mainFrame.modeFrame.text = SilentRotate.mainFrame.modeFrame:CreateFontString(nil, "ARTWORK")
-            SilentRotate.mainFrame.modeFrame.text:SetPoint("LEFT",2,0)
-            SilentRotate.mainFrame.modeFrame.text:SetTextColor(1,1,1,1)
-            SilentRotate.mainFrame.modeFrame.text:SetFont("Fonts\\ARIALN.ttf", 11)
-            SilentRotate.mainFrame.modeFrame.text:SetShadowColor(0,0,0,0.5)
-            SilentRotate.mainFrame.modeFrame.text:SetShadowOffset(1,-1)
+        if (not modeFrameText) then
+            modeFrameText = SilentRotate.mainFrame.modeFrame:CreateFontString(nil, "ARTWORK")
+            modeFrameText:SetPoint("LEFT",2,0)
+            modeFrameText:SetTextColor(1,1,1,1)
+            modeFrameText:SetFont("Fonts\\ARIALN.ttf", 11)
+            modeFrameText:SetShadowColor(0,0,0,0.5)
+            modeFrameText:SetShadowOffset(1,-1)
+            SilentRotate.mainFrame.modeFrame.text = modeFrameText
         else
-            SilentRotate.mainFrame.modeFrame.text:Show()
+            modeFrameText:Show()
         end
-        SilentRotate.mainFrame.modeFrame.text:SetText(L["NO_MODE_AVAILABLE"])
+        modeFrameText:SetText(L["NO_MODE_AVAILABLE"])
         for _, mapping in ipairs(modeFrameMapping) do
             local modeName = mapping.modeName
             SilentRotate.mainFrame.modeFrames[modeName]:Hide()
@@ -207,8 +273,8 @@ function SilentRotate:applyModeFrameSettings()
         return
     end
 
-    if SilentRotate.mainFrame.modeFrame.text then
-        SilentRotate.mainFrame.modeFrame.text:Hide()
+    if modeFrameText then
+        modeFrameText:Hide()
     end
 
     local commonModeWidth = SilentRotate.db.profile.mainFrameWidth/nbButtonsVisible
@@ -235,38 +301,53 @@ function SilentRotate:applyModeFrameSettings()
     end
 end
 
--- Create rotation frame
-function SilentRotate:createRotationFrame()
-    SilentRotate.mainFrame.rotationFrame = CreateFrame("Frame", 'rotationFrame', SilentRotate.mainFrame)
-    SilentRotate.mainFrame.rotationFrame:SetPoint('LEFT')
-    SilentRotate.mainFrame.rotationFrame:SetPoint('RIGHT')
-    SilentRotate.mainFrame.rotationFrame:SetPoint('TOP', 0, -SilentRotate.constants.titleBarHeight-SilentRotate.constants.modeBarHeight)
-    SilentRotate.mainFrame.rotationFrame:SetHeight(SilentRotate.constants.rotationFramesBaseHeight)
+-- Create background frame
+function SilentRotate:createBackgroundFrame(baseFrame, offsetY, height, frameName)
+    if not frameName then frameName = 'backgroundFrame' end
 
-    SilentRotate.mainFrame.rotationFrame.texture = SilentRotate.mainFrame.rotationFrame:CreateTexture(nil, "BACKGROUND")
-    SilentRotate.mainFrame.rotationFrame.texture:SetColorTexture(0,0,0,0.5)
-    SilentRotate.mainFrame.rotationFrame.texture:SetAllPoints()
+    backgroundFrame = CreateFrame("Frame", frameName, baseFrame)
+    backgroundFrame:SetPoint('LEFT')
+    backgroundFrame:SetPoint('RIGHT')
+    backgroundFrame:SetPoint('TOP', 0, -offsetY)
+    backgroundFrame:SetHeight(height)
+
+    backgroundFrame.texture = backgroundFrame:CreateTexture(nil, "BACKGROUND")
+    backgroundFrame.texture:SetColorTexture(0,0,0,0.5)
+    backgroundFrame.texture:SetAllPoints()
+
+    baseFrame[frameName] = backgroundFrame
+    return backgroundFrame
+end
+
+-- Create rotation frame
+function SilentRotate:createRotationFrame(baseFrame)
+    local offsetY = SilentRotate.constants.titleBarHeight+SilentRotate.constants.modeBarHeight
+    local height = SilentRotate.constants.rotationFramesBaseHeight
+    return SilentRotate:createBackgroundFrame(baseFrame, offsetY, height, 'rotationFrame')
 end
 
 -- Create backup frame
-function SilentRotate:createBackupFrame()
+function SilentRotate:createBackupFrame(baseFrame, rotationFrame)
     -- Backup frame
-    SilentRotate.mainFrame.backupFrame = CreateFrame("Frame", 'backupFrame', SilentRotate.mainFrame)
-    SilentRotate.mainFrame.backupFrame:SetPoint('TOPLEFT', SilentRotate.mainFrame.rotationFrame, 'BOTTOMLEFT', 0, 0)
-    SilentRotate.mainFrame.backupFrame:SetPoint('TOPRIGHT', SilentRotate.mainFrame.rotationFrame, 'BOTTOMRIGHT', 0, 0)
-    SilentRotate.mainFrame.backupFrame:SetHeight(SilentRotate.constants.rotationFramesBaseHeight)
+    backupFrame = CreateFrame("Frame", 'backupFrame', baseFrame)
+    backupFrame:SetPoint('TOPLEFT', rotationFrame, 'BOTTOMLEFT', 0, 0)
+    backupFrame:SetPoint('TOPRIGHT', rotationFrame, 'BOTTOMRIGHT', 0, 0)
+    backupFrame:SetHeight(SilentRotate.constants.rotationFramesBaseHeight)
 
     -- Set Texture
-    SilentRotate.mainFrame.backupFrame.texture = SilentRotate.mainFrame.backupFrame:CreateTexture(nil, "BACKGROUND")
-    SilentRotate.mainFrame.backupFrame.texture:SetColorTexture(0,0,0,0.5)
-    SilentRotate.mainFrame.backupFrame.texture:SetAllPoints()
+    backupFrame.texture = backupFrame:CreateTexture(nil, "BACKGROUND")
+    backupFrame.texture:SetColorTexture(0,0,0,0.5)
+    backupFrame.texture:SetAllPoints()
 
     -- Visual separator
-    SilentRotate.mainFrame.backupFrame.texture = SilentRotate.mainFrame.backupFrame:CreateTexture(nil, "BACKGROUND")
-    SilentRotate.mainFrame.backupFrame.texture:SetColorTexture(0.8,0.8,0.8,0.8)
-    SilentRotate.mainFrame.backupFrame.texture:SetHeight(1)
-    SilentRotate.mainFrame.backupFrame.texture:SetWidth(60)
-    SilentRotate.mainFrame.backupFrame.texture:SetPoint('TOP')
+    backupFrame.texture = backupFrame:CreateTexture(nil, "BACKGROUND")
+    backupFrame.texture:SetColorTexture(0.8,0.8,0.8,0.8)
+    backupFrame.texture:SetHeight(1)
+    backupFrame.texture:SetWidth(60)
+    backupFrame.texture:SetPoint('TOP')
+
+    baseFrame.backupFrame = backupFrame
+    return backupFrame
 end
 
 -- Create single hunter frame